simplify jump generation in auto-grow mode
All jumps are performed relative to the current position, and the
encoding is computed based upon jmpAddr - codeOffset. Which means
that the encoding is unaffected by the absolute position of the
code in memory. Thus, it is not necessary to treat the auto-grow
mode differently, the encoding can always be computed eagerly.